refactor: reorganizaed path for components
