--- frontend/src/app/(dashboard)/dashboard/student/enroll/[month]/date/payment/complete/page.tsx
@@ -1,71 +1,0 @@
-'use client'
-import React from 'react';
-import { useRouter } from 'next/navigation';
-import { ChevronLeftIcon } from '@heroicons/react/24/outline';
-
-export default function PaymentCompletePage() {
-  const router = useRouter();
-
-  const handleConfirm = () => {
-    // localStorage에서 선택된 세션 정보 삭제
-    if (typeof window !== 'undefined') {
-      localStorage.removeItem('selectedSessions');
-      localStorage.removeItem('selectedClassCards');
-    }
-    // 대시보드로 이동
-    router.push('/dashboard/student');
-  };
-
-  return (
-    <div className="flex flex-col min-h-screen bg-white font-[Pretendard Variable]">
-      {/* 헤더 */}
-      <header className="sticky top-0 z-40 flex flex-col bg-white border-b border-gray-200">
-        <div className="flex items-center justify-between px-4 h-14">
-          <button 
-            onClick={() => router.back()}
-            className="p-2 text-gray-600 hover:text-gray-900"
-          >
-            <ChevronLeftIcon className="w-5 h-5" />
-          </button>
-          <h1 className="absolute left-1/2 transform -translate-x-1/2 text-lg font-semibold text-gray-900">
-            수강신청
-          </h1>
-          <div className="w-10" />
-        </div>
-      </header>
-
-      {/* 완료 콘텐츠 */}
-      <div className="flex-1 flex flex-col items-center justify-center px-4 py-8">
-        <div className="flex flex-col items-center gap-6 w-full max-w-md">
-          {/* 완료 아이콘 */}
-          <div className="w-24 h-24 flex items-center justify-center">
-            <img src="/icons/complete.svg" alt="완료" className="w-full h-full" />
-          </div>
-
-          {/* 완료 메시지 */}
-          <div className="text-center">
-            <h2 className="text-xl font-semibold mb-2" style={{ color: '#573B30' }}>
-              수강신청이 완료되었어요!
-            </h2>
-            <p className="leading-relaxed" style={{ color: '#595959' }}>
-              입금 내역이 확인되면,<br />
-              수강신청이 최종 승인되어요
-            </p>
-          </div>
-        </div>
-      </div>
-
-      {/* 푸터 */}
-      <footer className="sticky bottom-0 flex flex-col w-full bg-white z-50">
-        <div className="flex gap-3 justify-center px-5 pt-2.5 pb-4 w-full text-base font-semibold leading-snug text-white">
-          <button
-            onClick={handleConfirm}
-            className="flex-1 shrink self-stretch px-2.5 py-4 rounded-lg min-w-[240px] size-full transition-colors duration-300 text-center bg-[#AC9592] text-white cursor-pointer hover:bg-[#9A8582]"
-          >
-            확인
-          </button>
-        </div>
-      </footer>
-    </div>
-  );
-}
