+++ frontend/src/app/(dashboard)/dashboard/student/enroll/page.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/navigation'
 import { useDashboardNavigation } from '@/contexts/DashboardContext'
-import { EnrollmentMainStep } from '@/components/dashboard/student/EnrollmentMainStep'
+import { EnrollmentMainStep } from '@/components/dashboard/student/Enrollment/EnrollmentMainStep'
 
 export default function EnrollPage() {
   const router = useRouter()
@@ -24,9 +24,11 @@ export default function EnrollPage() {
     )
   }
 
+  // SubPage가 설정되어 있으면 EnrollmentContainer가 렌더링되고,
+  // 그렇지 않으면 기본 수강신청 페이지를 렌더링
   if (subPage) {
-    return null 
+    return null // DashboardPage에서 EnrollmentContainer가 렌더링됨
   }
 
   return <EnrollmentMainStep />
-}
+}
