+++ frontend/src/app/(dashboard)/dashboard/student/profile/page.tsx
@@ -173,4 +173,4 @@ export default function ProfilePage() {
       )}
     </div>
   )
-}
+}
