+++ frontend/src/components/dashboard/student/Enrollment/enroll/EnrollmentDateStep.tsx
@@ -1,9 +1,9 @@
 'use client'
 import * as React from 'react'
-import Calendar from '@/app/(dashboard)/dashboard/student/enroll/[month]/date/Calendar'
-import DateSelectFooter from '@/app/(dashboard)/dashboard/student/enroll/[month]/date/DateSelectFooter'
+import Calendar from '@/components/features/student/enrollment/month/date/Calendar'
+import DateSelectFooter from '@/components/features/student/enrollment/month/date/DateSelectFooter'
 import { useState } from 'react'
-import { StatusStep } from '@/app/(dashboard)/dashboard/student/enroll/[month]/StatusStep'
+import { StatusStep } from '@/components/features/student/enrollment/month/StatusStep'
 import { ChevronLeftIcon } from '@heroicons/react/24/outline'
 import { useDashboardNavigation } from '@/contexts/DashboardContext'
 import { useEnrollmentCalculation } from '@/hooks/useEnrollmentCalculation'
