+++ frontend/src/components/features/student/classes/ClassDetailModal.tsx
