+++ frontend/src/components/features/student/enrollment/CheckboxAgreement.tsx
