+++ frontend/src/components/features/student/enrollment/EnrollmentCard.tsx
