+++ frontend/src/components/features/student/enrollment/EnrollmentPage.tsx
