+++ frontend/src/components/features/student/enrollment/NavigationItem.tsx
