+++ frontend/src/components/features/student/enrollment/Notice.tsx
