+++ frontend/src/components/features/student/enrollment/PolicySection.tsx
