+++ frontend/src/components/features/student/enrollment/StatusBadge.tsx
