+++ frontend/src/components/features/student/enrollment/month/ClassCard.tsx
