+++ frontend/src/components/features/student/enrollment/month/ClassDetails.tsx
