+++ frontend/src/components/features/student/enrollment/month/LocationSection.tsx
