+++ frontend/src/components/features/student/enrollment/month/SlideMessage.tsx
