+++ frontend/src/components/features/student/enrollment/month/date/payment/BankInfo.tsx
