--- frontend/src/components/student/SessionModal.tsx
@@ -1,140 +1,0 @@
-'use client'
-
-import React, { useState } from 'react'
-import { SlideUpModal } from '@/components/common/SlideUpModal'
-import { ClassDetailModal } from './ClassDetailModal'
-
-interface Session {
-  id: number
-  classId: number
-  date: string
-  startTime: string
-  endTime: string
-  class: {
-    id: number
-    className: string
-  }
-  session_id: number
-  enrollment_status: string
-  enrollment_id: number
-}
-
-// API 응답 타입과 호환되도록 any 타입도 허용
-type SessionData = Session | any
-
-interface ClassData {
-  id: number
-  className: string
-  dayOfWeek: string
-  startTime: string
-  teacher: {
-    id: number
-    name: string
-  }
-}
-
-interface SessionModalProps {
-  isOpen: boolean
-  selectedClass: ClassData | null
-  sessions: SessionData[]
-  onClose: () => void
-}
-
-export function SessionModal({ isOpen, selectedClass, sessions, onClose }: SessionModalProps) {
-  const [selectedClassId, setSelectedClassId] = useState<number | null>(null)
-  const [selectedClassSessions, setSelectedClassSessions] = useState<SessionData[]>([])
-  const [isClassDetailOpen, setIsClassDetailOpen] = useState(false)
-
-  if (!isOpen || !selectedClass) return null
-
-  const filteredSessions = sessions.filter((session: any) => session.class.id === selectedClass.id)
-
-  const handleSessionClick = (session: any) => {
-    // 해당 클래스의 모든 세션 정보를 함께 전달
-    console.log('세션 클릭됨:', session);
-    console.log('클래스 ID:', session.class.id);
-    
-    const classSessions = sessions.filter((s: any) => s.class.id === session.class.id);
-    console.log('해당 클래스의 모든 세션:', classSessions);
-    
-    setSelectedClassId(session.class.id);
-    setSelectedClassSessions(classSessions);
-    setIsClassDetailOpen(true);
-  }
-
-  const handleClassDetailClose = () => {
-    setIsClassDetailOpen(false)
-    setSelectedClassId(null)
-    setSelectedClassSessions([])
-  }
-
-  return (
-    <>
-      <SlideUpModal
-        isOpen={isOpen}
-        onClose={onClose}
-        title={`${selectedClass.className} - 세션 목록`}
-      >
-        <div className="pb-6">
-          {filteredSessions.length > 0 ? (
-            filteredSessions.map((session: any) => (
-              <div
-                key={`${session.session_id}-${session.enrollment_id}`}
-                className="p-4 mb-3 bg-blue-50 rounded-lg border border-blue-100 cursor-pointer hover:bg-blue-100 transition-colors"
-                onClick={() => handleSessionClick(session)}
-              >
-                <div className="flex justify-between items-start">
-                  <div>
-                    <p className="font-medium text-stone-700">
-                      {new Date(session.date).toLocaleDateString('ko-KR', {
-                        year: 'numeric',
-                        month: 'long',
-                        day: 'numeric',
-                        weekday: 'long'
-                      })}
-                    </p>
-                    <p className="text-sm text-stone-500 mt-1">
-                      {new Date(session.startTime).toLocaleTimeString('ko-KR', {
-                        hour: '2-digit',
-                        minute: '2-digit'
-                      })} - {new Date(session.endTime).toLocaleTimeString('ko-KR', {
-                        hour: '2-digit',
-                        minute: '2-digit'
-                      })}
-                    </p>
-                  </div>
-                  <span className={`px-2 py-1 text-xs rounded-full ${
-                    session.enrollment_status === 'PENDING' 
-                      ? 'bg-yellow-100 text-yellow-800' 
-                      : session.enrollment_status === 'CONFIRMED'
-                      ? 'bg-green-100 text-green-800'
-                      : 'bg-gray-100 text-gray-800'
-                  }`}>
-                    {session.enrollment_status}
-                  </span>
-                </div>
-              </div>
-            ))
-          ) : (
-            <div className="text-center py-8 text-stone-500">
-              <p>등록된 세션이 없습니다.</p>
-            </div>
-          )}
-        </div>
-      </SlideUpModal>
-
-      {/* 클래스 상세 정보 모달 */}
-      {selectedClassId && (
-        <>
-          {console.log('ClassDetailModal 렌더링 시도:', { selectedClassId, isClassDetailOpen, selectedClassSessions })}
-          <ClassDetailModal
-            isOpen={isClassDetailOpen}
-            onClose={handleClassDetailClose}
-            classId={selectedClassId}
-            classSessions={selectedClassSessions}
-          />
-        </>
-      )}
-    </>
-  )
-}
